Excel, pulls appts from Outlook/saves as image
There are multiple tabs in this excel file. the main tab, an instructions tab, and a settings tab. The person running the program must have access to any calendars they are trying to pull appt data from. There is a bit of setup, but once done it shouldn't have to be adjusted. You can even set the location to save the file to a sharepoint location so any time the file is updated it updates as an image on sharepoint as well.